requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
oauth2client==4.1.3
reportlab==4.0.7
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.3
openpyxl==3.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0